moved numba kernel
